:bug: fix: text area
