+++ views/airdrop/airdrop-form/airdrop-input.tsx
@@ -1,11 +1,11 @@
-import { Box, Button } from '@interest-protocol/ui-kit';
-import BigNumber from 'bignumber.js';
-import { FC, useMemo, useState } from 'react';
+import { Box } from '@interest-protocol/ui-kit';
+import { FC, useState } from 'react';
 import { useFormContext, useWatch } from 'react-hook-form';
 
 import { AirdropInputProps, IAirdropForm } from '../airdrop.types';
 import AirdropCustomAmountMethod from './airdrop-custom-amount-method';
 import AirdropNftCoinsMethod from './airdrop-nft-coins-method';
+import AirdropPreviewButton from './airdrop-preview-button';
 import AirdropPreviewModal from './airdrop-preview-modal';
 import AirdropUploadFile from './airdrop-upload-file';
 
@@ -14,23 +14,6 @@ const AirdropInput: FC<AirdropInputProps> = ({ setIsProgressView }) => {
   const token = useWatch({ control, name: 'token' });
   const method = useWatch({ control, name: 'method' });
   const [isSummaryOpen, setIsSummaryOpen] = useState(false);
-  const airdropList = useWatch({ control, name: 'airdropList' });
-
-  const isDisabled = useMemo(
-    () =>
-      !airdropList ||
-      !token?.balance ||
-      BigNumber(token.balance).lt(
-        BigNumber(
-          airdropList?.reduce(
-            (acc, { amount }) => acc.plus(BigNumber(amount ?? 0)),
-            BigNumber(0)
-          )
-        )
-      ),
-    [airdropList, token]
-  );
-
   const handleCloseSummaryModal = () => setIsSummaryOpen(false);
 
   const handleOpenSummaryModal = () => setIsSummaryOpen(true);
@@ -56,16 +39,7 @@ const AirdropInput: FC<AirdropInputProps> = ({ setIsProgressView }) => {
         onClose={handleCloseSummaryModal}
         setIsProgressView={setIsProgressView}
       />
-      <Box display="flex" justifyContent="center">
-        <Button
-          variant="filled"
-          borderRadius="xs"
-          disabled={isDisabled}
-          onClick={handleOpenSummaryModal}
-        >
-          Review & Confirm
-        </Button>
-      </Box>
+      <AirdropPreviewButton handleOpenSummaryModal={handleOpenSummaryModal} />
     </Box>
   );
 };
